fix(*): extracted auth logic into separate class, used axios instead of fetch
